pyproject.toml
[project]
name = "ditainer"
version = "1.1.0"
description = "Dependency Injection Container for Python"
authors = [
    { name = "Damian Ontivero", email = "damian.ontivero.dev@gmail.com" },
]
dependencies = ["pyyaml>=6.0.1"]
requires-python = "==3.12.*"
readme = "README.md"
license = { text = "MIT" }

[project.urls]
Homepage = "https://github.com/damian-ontivero/ditainer"
Repository = "https://github.com/damian-ontivero/ditainer"

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.pdm]
distribution = true

[tool.pdm.build]
package-dir = "src"
includes = ["src/ditainer"]
excludes = ["./**/.git"]
source-includes = ["tests", "LICENSE", "README.md"]

[tool.pdm.dev-dependencies]
dev = ["pre-commit>=3.8.0", "ipdb>=0.13.13"]
test = ["pytest>=8.3.2"]

[tool.pdm.scripts]
pc-install = { cmd = "pre-commit install" }
pc-run = { cmd = "pre-commit run --all-files" }
lint = { cmd = "pre-commit run ruff --all-files" }
tests = { cmd = "pytest tests -vs" }


[tool.ruff]
src = ["src"]
line-length = 120

[tool.ruff.format]
skip-magic-trailing-comma = true

[tool.ruff.lint]
select = [
    "F", # Pyflakes
    "I", # isort
]

[tool.ruff.lint.isort]
known-first-party = ["ditainer", "tests"]
force-single-line = true
lines-between-types = 1
lines-after-imports = 2
split-on-trailing-comma = false
